requirements.txt
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
discord-protos==0.0.2
discord.py-self==2.1.0
Flask==3.1.3
frozenlist==1.8.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pytz==2025.2
multidict==6.7.1
protobuf==6.33.5
Pygments==2.19.2
requests==2.32.5
rich==14.3.3
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.6
yarl==1.22.0
aiosqlite==0.22.1
